feat: rename files that aren't meant to be wired directly for SDC
