+++ src/components/forms/forms.stories.js
@@ -3,7 +3,7 @@ import radio from './radio/radio.twig';
 import select from './select/select.twig';
 import textfields from './textfields/textfields.twig';
 
-import { props } from './forms.component.yml';
+import { props } from './forms.yml';
 
 const checkboxData = props.properties.checkboxes.data;
 const radioData = props.properties.radios.data;
