+++ src/components/navigation/pager/pager-both-ellipses.yml
