+++ src/components/navigation/pager/pager-next-ellipses.yml
