+++ src/components/navigation/pager/pager.stories.js
@@ -2,10 +2,10 @@ import pagerTwig from './pager.twig';
 import { mapDataToTwig } from '../../../util/dataTransformers.js';
 
 import { props as pagerProps } from './pager.component.yml';
-import { props as pagerNextEllipsesProps } from './pager-next-ellipses.component.yml';
-import { props as pagerPrevEllipsesProps } from './pager-prev-ellipses.component.yml';
-import { props as pagerBothEllipsesProps } from './pager-both-ellipses.component.yml';
-import { props as pagerFirstLastProps } from './pager-first-last.component.yml';
+import { props as pagerNextEllipsesProps } from './pager-next-ellipses.yml';
+import { props as pagerPrevEllipsesProps } from './pager-prev-ellipses.yml';
+import { props as pagerBothEllipsesProps } from './pager-both-ellipses.yml';
+import { props as pagerFirstLastProps } from './pager-first-last.yml';
 
 const pagerData = mapDataToTwig(pagerProps.properties);
 const pagerNextEllipsesData = mapDataToTwig(pagerNextEllipsesProps.properties);
